Generalization function working
Generalization function has been fixed and now works properly for all types of data, be it ranges, strings and number. A new function has been added to remove white spaces from the dataframe column names in order to avoid issues with the dictionaries. New parameters added to the DataFly function to add hierarchies to the generalization.